% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readStudyShinySelection.R
\name{readStudyShinySelection}
\alias{readStudyShinySelection}
\title{Read areas, links, clusters & districts selection from file}
\usage{
readStudyShinySelection(input_path)
}
\arguments{
\item{input_path}{\code{character}. Path of .xlsx input file}
}
\value{
\code{list}
}
\description{
Read areas, links, clusters & districts selection from file
}
